+++ docs/api/IO/Libraries/mesh/index.rst
@@ -5,7 +5,7 @@ Mesh Reader
 
 
 The below functions are helper functions for the Fortran binary mesh database
-reader. The functions are called in :cpp:func:`specfem::IO::read_mesh`
+reader. The functions are called in :cpp:func:`specfem::IO::read_2d_mesh`
 in sequential order and faciliate the reading of the mesh database.
 
 .. doxygenfunction:: specfem::IO::mesh::impl::fortran::read_mesh_database_header
