+++ docs/api/IO/index.rst
@@ -3,7 +3,7 @@ Input/Output modules
 ====================
 
 There are three high-level functions that are essential for the SPECFEM++
-codebase: :cpp:func:`specfem::IO::read_mesh`,
+codebase: :cpp:func:`specfem::IO::read_2d_mesh`,
 :cpp:func:`specfem::IO::read_sources`, and
 :cpp:func:`specfem::IO::read_receivers`. These functions are
 used to read the mesh, sources and receivers from disk. The underlying
@@ -32,7 +32,7 @@ are exposed through the following modules:
 Read Mesh
 ---------
 
-.. doxygenfunction:: specfem::IO::read_mesh
+.. doxygenfunction:: specfem::IO::read_2d_mesh
 
 
 Read Sources
