+++ src/IO/mesh/impl/fortran/dim2/read_material_properties.cpp
@@ -1,4 +1,4 @@
-#include "IO/mesh/impl/fortran/read_material_properties.hpp"
+#include "IO/mesh/impl/fortran/dim2/read_material_properties.hpp"
 #include "IO/fortranio/interface.hpp"
 #include "enumerations/interface.hpp"
 // #include "mesh/materials/materials.hpp"
