+++ src/IO/mesh/impl/fortran/dim2/read_parameters.cpp
@@ -1,4 +1,4 @@
-#include "IO/mesh/impl/fortran/read_parameters.hpp"
+#include "IO/mesh/impl/fortran/dim2/read_parameters.hpp"
 #include "IO/fortranio/interface.hpp"
 #include "mesh/parameters/parameters.hpp"
