+++ src/execute.cpp
@@ -42,7 +42,7 @@ void execute(
   //                   Read mesh and materials
   // --------------------------------------------------------------
   const auto quadrature = setup.instantiate_quadrature();
-  const auto mesh = specfem::IO::read_mesh(database_filename, mpi);
+  const auto mesh = specfem::IO::read_2d_mesh(database_filename, mpi);
   // --------------------------------------------------------------
 
   // --------------------------------------------------------------
