+++ tests/unit-tests/algorithms/interpolate_function.cpp
@@ -21,7 +21,7 @@ TEST(ALGORITHMS, interpolate_function) {
 
   // Read Mesh database
   specfem::MPI::MPI *mpi = MPIEnvironment::get_mpi();
-  specfem::mesh::mesh mesh = specfem::IO::read_mesh(database_file, mpi);
+  specfem::mesh::mesh mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
   constexpr int N = 5;
