+++ tests/unit-tests/algorithms/locate.cpp
@@ -14,7 +14,7 @@ TEST(ALGORITHMS, locate_point) {
 
   // Read Mesh database
   specfem::MPI::MPI *mpi = MPIEnvironment::get_mpi();
-  specfem::mesh::mesh mesh = specfem::IO::read_mesh(database_file, mpi);
+  specfem::mesh::mesh mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
   // Quadratures
   specfem::quadrature::gll::gll gll(0.0, 0.0, 5);
