+++ tests/unit-tests/assembly/test_fixture/test_fixture.cpp
@@ -32,7 +32,7 @@ ASSEMBLY::ASSEMBLY() {
   for (auto &Test : Tests) {
     const auto [database_file, sources_file, stations_file] =
         Test.get_databases();
-    const auto mesh = specfem::IO::read_mesh(database_file, mpi);
+    const auto mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
     this->Meshes.push_back(mesh);
