+++ tests/unit-tests/compute/coupled_interfaces/coupled_interfaces_tests.cpp
@@ -178,7 +178,7 @@ TEST(COMPUTE_TESTS, coupled_interfaces_tests) {
 
     // Read mesh generated MESHFEM
     specfem::mesh::mesh mesh =
-        specfem::IO::read_mesh(Test.databases.mesh.database_filename, mpi);
+        specfem::IO::read_2d_mesh(Test.databases.mesh.database_filename, mpi);
 
     // Generate compute structs to be used by the solver
     specfem::compute::mesh assembly(mesh.control_nodes, quadratures);
