+++ tests/unit-tests/displacement_tests/Newmark/elastic/newmark_tests.cpp
@@ -64,7 +64,7 @@ TEST(DISPLACEMENT_TESTS, newmark_scheme_tests) {
 
   // Read mesh generated MESHFEM
   std::vector<specfem::medium::material *> materials;
-  specfem::mesh::mesh mesh = specfem::IO::read_mesh(database_file, mpi);
+  specfem::mesh::mesh mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
   // Read sources
   //    if start time is not explicitly specified then t0 is determined using
