+++ tests/unit-tests/domain/elastic/rmass_inverse_tests.cpp
@@ -61,7 +61,7 @@ TEST(DOMAIN_TESTS, rmass_inverse_elastic_test) {
 
   // Read mesh generated MESHFEM
   std::vector<specfem::medium::material *> materials;
-  specfem::mesh::mesh mesh = specfem::IO::read_mesh(database_file, mpi);
+  specfem::mesh::mesh mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
   // Read sources
   //    if start time is not explicitly specified then t0 is determined using
