+++ tests/unit-tests/domain/rmass_inverse_tests.cpp
@@ -117,7 +117,7 @@ TEST(DOMAIN_TESTS, rmass_inverse) {
     std::cout << "Reading mesh file: " << database_file << std::endl;
 
     // Read mesh generated MESHFEM
-    specfem::mesh::mesh mesh = specfem::IO::read_mesh(database_file, mpi);
+    specfem::mesh::mesh mesh = specfem::IO::read_2d_mesh(database_file, mpi);
 
     std::cout << "Setting up sources and receivers" << std::endl;
